requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper>=20231117
transformers>=4.35.0
sentencepiece>=0.1.99
protobuf>=3.20.0
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
